Add the ability to examine item 'c'ontents on inventory screens
For some items, such as sealed cans of food, the current 'e'xamine
behaviour on inventory screens results in a great deal of information
about the container at the top of the page, when in fact, the player
was interested in the contents of that container.  With the addition
of this functionality, the player will be able to quickly examine the
contents of items.